ConvRF 3-Class Classification Experiment
Reference issue
Issue #99 Convolutional Forests

Type of change
Multi-class classification experiments and Kappa plots for ConvRF
